requirements.txt
snowflake-connector-python[pandas]
python-dotenv
pandas
pyarrow
fastparquet
scikit-learn
catboost
lightgbm
xgboost
streamlit
plotly
matplotlib
seaborn
